intellidocs/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pdfplumber==0.9.0
python-docx==0.8.11
sentence-transformers==2.2.2
faiss-cpu==1.7.4
google-generativeai==0.3.2
numpy==1.24.3
gunicorn==21.2.0